app/requirements.txt
numpy
streamlit
tensorflow
pillow

